9.2. Checkstyle [#439573]

+++ src/ru/job4j/condition/SqArea.java
@@ -2,7 +2,7 @@ package ru.job4j.condition;
 
 public class SqArea {
     public static double square(int p, int k) {
-        return k * (p / (2 * (k+1)));
+        return k * ((float) p / (2 * (float) k + 1));
     }
 
     public static void main(String[] args) {
